+++ src/scriptlets/scriptlets-list.js
@@ -55,6 +55,7 @@ export * from './trusted-set-cookie';
 export * from './trusted-set-cookie-reload';
 export * from './trusted-replace-fetch-response';
 export * from './trusted-set-local-storage-item';
+export * from './trusted-set-session-storage-item';
 export * from './trusted-set-constant';
 export * from './inject-css-in-shadow-dom';
 export * from './remove-node-text';